Finishing Touches

+++ src/app/components/inventory-index/inventory-index.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { InventoryService } from '../../services/inventory.service';
+import { ItemService } from '../../services/item.service';
 import { Inventory } from '../../models/inventory.model';
-import {DatePipe, NgForOf} from "@angular/common";
+import { Item } from '../../models/item.model';
+import { DatePipe, NgForOf } from '@angular/common';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-inventory-index',
@@ -15,23 +18,50 @@ import {DatePipe, NgForOf} from "@angular/common";
 })
 export class InventoryIndexComponent implements OnInit {
   inventoryItems: Inventory[] = [];
+  items: Item[] = [];
 
-  constructor(private inventoryService: InventoryService) {}
+  constructor(
+    private inventoryService: InventoryService,
+    private itemService: ItemService,
+    private router: Router
+  ) {}
 
   ngOnInit(): void {
-    this.loadInventory();
+    this.loadItems();
+  }
+
+  loadItems(): void {
+    this.itemService.getItems().subscribe(items => {
+      this.items = items;
+      this.loadInventory();
+    });
+  }
+
+  editInventory(id: number): void {
+    this.router.navigate(['/inventory-update', id]);
+  }
+
+  createInventory(): void {
+    this.router.navigate(['/inventory-create']);
   }
 
   loadInventory(): void {
-    this.inventoryService.getInventory().subscribe(items => {
-      this.inventoryItems = items;
+    this.inventoryService.getInventory().subscribe(inventory => {
+      this.inventoryItems = inventory.map(inv => {
+        const item = this.items.find(i => i.item_Id === inv.item_Id);
+        return {
+          ...inv,
+          itemName: item ? item.item_Name : 'Unknown Item'
+        };
+      });
     });
   }
 
   deleteItem(id: number): void {
-    this.inventoryService.deleteInventoryItem(id).subscribe(() => {
-      // Reload inventory list after deletion
-      this.loadInventory();
-    });
+    if (confirm('Are you sure you want to delete this item?')) {
+      this.inventoryService.deleteInventoryItem(id).subscribe(() => {
+        this.inventoryItems = this.inventoryItems.filter(i => i.inventory_Id !== id);
+      });
+    }
   }
 }
